+++ cmd/zetae2e/local/local.go
@@ -188,7 +188,7 @@ func localE2ETest(cmd *cobra.Command, _ []string) {
 	if !skipSetup {
 		logger.Print("⚙️ setting up networks")
 		startTime := time.Now()
-		deployerRunner.SetupEVM(contractsDeployed)
+		deployerRunner.SetupEVM(contractsDeployed, true)
 		deployerRunner.SetZEVMContracts()
 
 		// NOTE: this method return an error so we handle it and panic if it occurs unlike other method that panics directly
@@ -250,7 +250,11 @@ func localE2ETest(cmd *cobra.Command, _ []string) {
 			e2etests.TestZetaDepositRestrictedName,
 		}
 		bitcoinTests := []string{
-			e2etests.TestBitcoinWithdrawName,
+			e2etests.TestBitcoinWithdrawSegWitName,
+			e2etests.TestBitcoinWithdrawTaprootName,
+			e2etests.TestBitcoinWithdrawLegacyName,
+			e2etests.TestBitcoinWithdrawP2SHName,
+			e2etests.TestBitcoinWithdrawP2WSHName,
 			e2etests.TestBitcoinWithdrawInvalidAddressName,
 			e2etests.TestZetaWithdrawBTCRevertName,
 			e2etests.TestCrosschainSwapName,
@@ -285,6 +289,12 @@ func localE2ETest(cmd *cobra.Command, _ []string) {
 			e2etests.TestPauseZRC20Name,
 			e2etests.TestUpdateBytecodeName,
 			e2etests.TestDepositEtherLiquidityCapName,
+
+			// TestMigrateChainSupportName tests EVM chain migration. Currently this test doesn't work with Anvil because pre-EIP1559 txs are not supported
+			// See issue below for details
+			// TODO: renenable this test as per the issue below
+			// https://github.com/zeta-chain/node/issues/1980
+			// e2etests.TestMigrateChainSupportName,
 		))
 	}
 	if testPerformance {